Move packagejson to its own analyzer for reusability
