+++ pkg/analysis/passes/packagejson/testdata/human-package-json/package.json
