+++ pkg/analysis/passes/packagejson/testdata/invalid-package-json/package.json
