+++ pkg/analysis/passes/packagejson/testdata/version-mismatch/package.json
