+++ pkg/analysis/passes/sourcecode/sourcecode_test.go
@@ -80,137 +80,3 @@ func TestSourceCodeNotProvidedReportAll(t *testing.T) {
 	require.Equal(t, "Source code not provided or the provided URL  does not point to a valid source code repository", interceptor.Diagnostics[0].Title)
 	require.Equal(t, nil, sourceCodeDir)
 }
-
-func TestVersionSourceCodeMatch(t *testing.T) {
-	var interceptor testpassinterceptor.TestPassInterceptor
-	pluginJsonContent := []byte(`{
-		"id": "` + pluginId + `",
-		"type": "panel",
-		"info": {
-			"version": "2.1.2"
-		}
-	}`)
-
-	pass := &analysis.Pass{
-		RootDir:       filepath.Join("./"),
-		SourceCodeDir: filepath.Join("testdata", "version-match"),
-		ResultOf: map[*analysis.Analyzer]interface{}{
-			metadata.Analyzer: pluginJsonContent,
-		},
-		Report: interceptor.ReportInterceptor(),
-	}
-
-	sourceCodeDir, err := Analyzer.Run(pass)
-	require.NoError(t, err)
-
-	require.Len(t, interceptor.Diagnostics, 0)
-	require.Equal(t, filepath.Join("testdata", "version-match"), sourceCodeDir)
-}
-
-func TestNoPackageJson(t *testing.T) {
-	var interceptor testpassinterceptor.TestPassInterceptor
-	pluginJsonContent := []byte(`{
-		"id": "` + pluginId + `",
-		"type": "panel",
-		"info": {
-			"version": "2.1.2"
-		}
-	}`)
-
-	pass := &analysis.Pass{
-		RootDir:       filepath.Join("./"),
-		SourceCodeDir: filepath.Join("testdata", "no-package-json"),
-		ResultOf: map[*analysis.Analyzer]interface{}{
-			metadata.Analyzer: pluginJsonContent,
-		},
-		Report: interceptor.ReportInterceptor(),
-	}
-
-	_, err := Analyzer.Run(pass)
-	require.NoError(t, err)
-
-	require.Len(t, interceptor.Diagnostics, 1)
-	require.Equal(t, "Could not find or parse package.json from testdata/no-package-json", interceptor.Diagnostics[0].Title)
-}
-
-func TestInvalidPackageJson(t *testing.T) {
-	var interceptor testpassinterceptor.TestPassInterceptor
-	pluginJsonContent := []byte(`{
-		"id": "` + pluginId + `",
-		"type": "panel",
-		"info": {
-			"version": "2.1.2"
-		}
-	}`)
-
-	pass := &analysis.Pass{
-		RootDir:       filepath.Join("./"),
-		SourceCodeDir: filepath.Join("testdata", "invalid-package-json"),
-		ResultOf: map[*analysis.Analyzer]interface{}{
-			metadata.Analyzer: pluginJsonContent,
-		},
-		Report: interceptor.ReportInterceptor(),
-	}
-
-	_, err := Analyzer.Run(pass)
-	require.NoError(t, err)
-
-	require.Len(t, interceptor.Diagnostics, 1)
-	require.Equal(t, "Could not find or parse package.json from testdata/invalid-package-json", interceptor.Diagnostics[0].Title)
-}
-
-// package.json in source code sometimes use a "human json" format that allows
-// for comments or trailing commas. This is not valid json, but we should still be able to parse it
-// because developers are used to this format.
-func TestAllowHumanJson(t *testing.T) {
-	var interceptor testpassinterceptor.TestPassInterceptor
-	pluginJsonContent := []byte(`{
-		"id": "` + pluginId + `",
-		"type": "panel",
-		"info": {
-			"version": "2.1.5"
-		}
-	}`)
-
-	pass := &analysis.Pass{
-		RootDir:       filepath.Join("./"),
-		SourceCodeDir: filepath.Join("testdata", "human-package-json"),
-		ResultOf: map[*analysis.Analyzer]interface{}{
-			metadata.Analyzer: pluginJsonContent,
-		},
-		Report: interceptor.ReportInterceptor(),
-	}
-
-	_, err := Analyzer.Run(pass)
-	require.NoError(t, err)
-
-	require.Len(t, interceptor.Diagnostics, 0)
-}
-
-func TestVersionMisMatch(t *testing.T) {
-	var interceptor testpassinterceptor.TestPassInterceptor
-
-	// note the test file has version 2.1.2
-	pluginJsonContent := []byte(`{
-		"id": "` + pluginId + `",
-		"type": "panel",
-		"info": {
-			"version": "2.1.5"
-		}
-	}`)
-
-	pass := &analysis.Pass{
-		RootDir:       filepath.Join("./"),
-		SourceCodeDir: filepath.Join("testdata", "version-mismatch"),
-		ResultOf: map[*analysis.Analyzer]interface{}{
-			metadata.Analyzer: pluginJsonContent,
-		},
-		Report: interceptor.ReportInterceptor(),
-	}
-
-	_, err := Analyzer.Run(pass)
-	require.NoError(t, err)
-
-	require.Len(t, interceptor.Diagnostics, 1)
-	require.Equal(t, "The version in package.json (2.1.2) doesn't match the version in plugin.json (2.1.5)", interceptor.Diagnostics[0].Title)
-}
